requirements.txt
fastapi==0.78.0
pydantic==1.9.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.20.0
SQLAlchemy==1.4.37
uvicorn==0.17.6
gunicorn==20.1.0
